requirements.txt
segmentation-models-pytorch
torch
PageXML-to-Mask-converter
torchvision
pillow
numpy
ttach
setuptools
matplotlib
pandas
seaborn
tqdm
albumentations
scikit-image
opencv-python
sklearn
ezodf
lxml
shapely
future
scikit-learn
dataclasses
dataclasses_json
scikit-learn
scipy
loguru
git+https://gitlab2.informatik.uni-wuerzburg.de/nof36hn/binarizer-doxa.git#egg=doxapy
pyserde